+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/artemis_context/ArtemisContext.kt
@@ -14,6 +14,9 @@ sealed interface ArtemisContext {
 
     sealed interface AccountLoaded: LoggedIn {
         val account: Account
+
+        val clientId: Long
+            get() = account.id
     }
 
     sealed interface Course: AccountLoaded {
