+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/artemis_context/impl/ArtemisContextImpl.kt
@@ -1,5 +1,6 @@
-package de.tum.informatics.www1.artemis.native_app.core.data.artemis_context
+package de.tum.informatics.www1.artemis.native_app.core.data.artemis_context.impl
 
+import de.tum.informatics.www1.artemis.native_app.core.data.artemis_context.ArtemisContext
 import de.tum.informatics.www1.artemis.native_app.core.model.account.Account
