+++ feature/metis/shared/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/shared/ui/profile_picture/UserProfileDialog.kt
@@ -62,8 +62,7 @@ fun UserProfileDialog(
         parametersOf(courseId, userId)
     }
 
-    val clientId = courseArtemisContext.account.id
-    val isSendMessageAvailable = clientId != userId
+    val isSendMessageAvailable = courseArtemisContext.clientId != userId
     val context = LocalContext.current
 
     UserProfileDialogImpl(
